requirements.txt
discord.py>=2.3.2
mysql-connector-python>=8.0.33
python-dotenv>=1.0.1
cryptography>=42.0.0
